pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Python page analyzer"
authors = ["MoiseenkoVA <mvaload@gmail.com>"]
packages = [
  {include = "page_analyzer"},
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
